# backend/public/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Train Booking API",
    "version": "1.0.0",
    "description": "Static OpenAPI spec for the Train Booking backend"
  },
  "servers": [{ "url": "/api", "description": "Local API base path" }],
  "paths": {
...
    "/routes": {
      "get": {
        "summary": "List routes",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["asc", "desc"] },
            "description": "Station order direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Routes retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "routes": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TrainRoute" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a route",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "Main Line" },
                  "description": { "type": "string", "example": "Primary railway route line" },
                  "is_active": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Route created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Route created successfully." },
                    "route": { "$ref": "#/components/schemas/TrainRoute" }
                  }
                }
              }
            }
          },
          "422": { "description": "Validation errors" }
        }
      }
    },
    "/routes/{route}": {
      "get": {
        "summary": "Get a route",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "route", "in": "path", "required": true, "schema": { "type": "integer" } },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["asc", "desc"] },
            "description": "Station order direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Route retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "route": { "$ref": "#/components/schemas/TrainRoute" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a route",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "route", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Main Line" },
                  "description": { "type": "string", "example": "Updated route details" },
                  "is_active": { "type": "boolean", "example": false }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Route updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Route updated successfully." },
                    "route": { "$ref": "#/components/schemas/TrainRoute" }
                  }
                }
              }
            }
          },
          "422": { "description": "Validation errors" }
        }
      },
      "delete": {
        "summary": "Delete a route",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "route", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Route deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Route deleted successfully." }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "TrainRoute": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Main Line" },
          "description": { "type": "string", "nullable": true, "example": "Primary railway route line" },
          "is_active": { "type": "boolean", "example": true },
          "created_by": { "type": "integer", "nullable": true, "example": 1 },
          "updated_by": { "type": "integer", "nullable": true, "example": 1 },
          "deleted_by": { "type": "integer", "nullable": true, "example": null },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true, "example": null },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "stations": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RouteStation" }
          }
        }
      },
      "RouteStation": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "route_id": { "type": "integer", "example": 1 },
          "station_id": { "type": "integer", "example": 2 },
          "sequence": { "type": "integer", "example": 1 },
          "station": { "$ref": "#/components/schemas/Station" },
          "created_by": { "type": "integer", "nullable": true, "example": 1 },
          "updated_by": { "type": "integer", "nullable": true, "example": 1 },
          "deleted_by": { "type": "integer", "nullable": true, "example": null },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true, "example": null },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Station": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Central Station" },
          "address": { "type": "string", "example": "123 Main Street, City" },
          "is_active": { "type": "boolean", "example": true },
          "created_by": { "type": "integer", "nullable": true, "example": 1 },
          "updated_by": { "type": "integer", "nullable": true, "example": 1 },
          "deleted_by": { "type": "integer", "nullable": true, "example": null },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true, "example": null },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}